Module pallet_tidefi_stake::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

GenesisConfig

Genesis configuration

Pallet

The pallet implementing the on-chain logic.

Enums

BatchType

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

Type Definitions

AccountStakes

Account staking by CurrencyID

InterestCompoundLastSession

The last session we should compound the account interests.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

OperatorAccountId

Operator account

PendingStoredSessions

Map from all pending stored sessions.

QueueCompound

QueueUnstake

SessionTotalFees

The total fees for the session. If total hasn’t been set or has been removed then 0 stake is returned.

StakingCurrencyMeta

Staking metadata defined by the council (minimum and maximum stake amount)

StakingPeriodRewards

Staking rewards defined by the council

StakingPool

Staking pool

UnstakeFee

The percentage of fee when unstake is done before the ending.