Struct pallet_tidefi_stake::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn stake(
origin: OriginFor<T>,
currency_id: CurrencyId,
amount: Balance,
duration: T::BlockNumber
) -> DispatchResultWithPostInfo

Stake currency

currency_id: The currency to stake

amount: The amount to stake

duration: The duration is in numbers of blocks. (blocks are ~6seconds)

Emits Staked event when successful.

Weight: O(1)

pub fn unstake(
origin: OriginFor<T>,
stake_id: Hash,
force_unstake: bool
) -> DispatchResultWithPostInfo

Unstake

stake_id: Unique Stake ID

force_unstake: Unstake with extra fees, even if the staking is not expired

Emits Unstaked event when successful.

Weight: O(1)

impl<T: Config> Pallet<T>

pub fn account_id() -> T::AccountId

pub fn add_account_stake(
account_id: &T::AccountId,
currency_id: CurrencyId,
amount: Balance,
duration: T::BlockNumber
) -> Result<Hash, DispatchError>

pub fn do_next_compound_interest_operation(
max_weight: Weight
) -> Result<(Weight, bool), DispatchError>

pub fn do_next_unstake_operation(
max_weight: Weight
) -> Result<(Weight, bool), DispatchError>

pub fn get_account_stakes(
account_id: &T::AccountId
) -> Vec<(CurrencyId, Stake<BalanceInfo, T::BlockNumber>)>

impl<T: Config> Pallet<T>

pub fn staking_pool<KArg>(k: KArg) -> Option<Balance> where
KArg: EncodeLike<CurrencyId>,

Staking pool

impl<T: Config> Pallet<T>

pub fn staking_rewards(
) -> BoundedVec<(T::BlockNumber, Percent), T::StakingRewardCap>

Staking rewards defined by the council

impl<T: Config> Pallet<T>

pub fn staking_meta<KArg>(k: KArg) -> Option<StakeCurrencyMeta<Balance>> where
KArg: EncodeLike<CurrencyId>,

Staking metadata defined by the council (minimum and maximum stake amount)

impl<T: Config> Pallet<T>

pub fn unstake_fee() -> Percent

The percentage of fee when unstake is done before the ending.

impl<T: Config> Pallet<T>

pub fn interest_compound_last_session() -> SessionIndex

The last session we should compound the account interests.

impl<T: Config> Pallet<T>

pub fn unstake_queue(
) -> BoundedVec<(T::AccountId, Hash, T::BlockNumber), T::UnstakeQueueCap>

Manage which we should pay off to.

impl<T: Config> Pallet<T>

pub fn pending_sessions<KArg>(k: KArg) -> Option<()> where
KArg: EncodeLike<SessionIndex>,

Map from all pending stored sessions.

impl<T: Config> Pallet<T>

pub fn session_total_fees<KArg1, KArg2>(k1: KArg1, k2: KArg2) -> Balance where
KArg1: EncodeLike<SessionIndex>,
KArg2: EncodeLike<CurrencyId>,

The total fees for the session. If total hasn’t been set or has been removed then 0 stake is returned.

impl<T: Config> Pallet<T>

pub fn account_stakes<KArg>(
k: KArg
) -> BoundedVec<Stake<Balance, T::BlockNumber>, T::StakeAccountCap> where
KArg: EncodeLike<T::AccountId>,

Account staking by CurrencyID