Trait pallet_tidefi_stake::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type StakePalletId: Get<PalletId>;
    type StakeAccountCap: Get<u32>;
    type StakingRewardCap: Get<u32>;
    type BlocksForceUnstake: Get<Self::BlockNumber>;
    type BatchSize: Get<u32>;
    type WeightInfo: WeightInfo;
    type AssetRegistry: AssetRegistryExt;
    type Security: SecurityExt<Self::AccountId, Self::BlockNumber>;
    type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Events

type StakePalletId: Get<PalletId>

Pallet ID

type StakeAccountCap: Get<u32>

Maximum active stake / account

type StakingRewardCap: Get<u32>

Maximum number of periods defined

type BlocksForceUnstake: Get<Self::BlockNumber>

Number of block to wait before unstake if forced.

type BatchSize: Get<u32>

Batch size.

This many accounts and unstake are processed in each on_idle` request.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type AssetRegistry: AssetRegistryExt

Asset registry traits

type Security: SecurityExt<Self::AccountId, Self::BlockNumber>

Security traits

type CurrencyTidefi: Inspect<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Mutate<Self::AccountId, AssetId = CurrencyId, Balance = Balance> + Transfer<Self::AccountId, AssetId = CurrencyId, Balance = Balance>

Tidechain currency wrapper